changes requested Removed DropShadowPanel.Members.cs file Moved all methods to DropShadowPanel.cs Renamed Loaded and SizeChanged event handlers to OnLoaded and OnSizeChanged Remove template binding of content in xaml null checks for _shadowVisual and _dropShadow
