--- Microsoft.Toolkit.Uwp.UI.Controls/DropShadowPanel/DropShadowPanel.Members.cs
@@ -1,155 +1,0 @@
-﻿// ******************************************************************
-// Copyright (c) Microsoft. All rights reserved.
-// This code is licensed under the MIT License (MIT).
-// THE CODE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
-// INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-// IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-// TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH
-// THE CODE OR THE USE OR OTHER DEALINGS IN THE CODE.
-// ******************************************************************
-
-using System.Numerics;
-using Windows.ApplicationModel;
-using Windows.Foundation.Metadata;
-using Windows.UI;
-using Windows.UI.Composition;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Shapes;
-
-namespace Microsoft.Toolkit.Uwp.UI.Controls
-{
-    /// <summary>
-    /// The <see cref="DropShadowPanel"/> control allows the creation of a DropShadow for any Xaml FrameworkElement in markup
-    /// making it easier to add shadows to Xaml without having to directly drop down to Windows.UI.Composition APIs.
-    /// </summary>
-    public partial class DropShadowPanel
-    {
-        /// <summary>
-        /// Gets a value indicating whether the platform supports drop shadows.
-        /// </summary>
-        /// <remarks>
-        /// On platforms not supporting drop shadows, this control has no effect.
-        /// </remarks>
-        public static bool IsSupported =>
-            !DesignMode.DesignModeEnabled &&
-            ApiInformation.IsTypePresent("Windows.UI.Composition.DropShadow"); // SDK >= 14393
-
-        /// <summary>
-        /// Gets DropShadow. Exposes the underlying composition object to allow custom Windows.UI.Composition animations.
-        /// </summary>
-        public DropShadow DropShadow => _dropShadow;
-
-        /// <summary>
-        /// Gets or sets the mask of the underlying <see cref="Windows.UI.Composition.DropShadow"/>.
-        /// Allows for a custom <see cref="Windows.UI.Composition.CompositionBrush"/> to be set.
-        /// </summary>
-        public CompositionBrush Mask
-        {
-            get
-            {
-                return _dropShadow.Mask;
-            }
-
-            set
-            {
-                _dropShadow.Mask = value;
-            }
-        }
-
-        private void CompositionShadow_Loaded(object sender, RoutedEventArgs e)
-        {
-            ConfigureShadowVisualForCastingElement();
-        }
-
-        private void CompositionShadow_SizeChanged(object sender, SizeChangedEventArgs e)
-        {
-            UpdateShadowSize();
-        }
-
-        private void ConfigureShadowVisualForCastingElement()
-        {
-            UpdateShadowMask();
-
-            UpdateShadowSize();
-        }
-
-        private void OnBlurRadiusChanged(double newValue)
-        {
-            _dropShadow.BlurRadius = (float)newValue;
-        }
-
-        private void OnColorChanged(Color newValue)
-        {
-            _dropShadow.Color = newValue;
-        }
-
-        private void OnOffsetXChanged(double newValue)
-        {
-            UpdateShadowOffset((float)newValue, _dropShadow.Offset.Y, _dropShadow.Offset.Z);
-        }
-
-        private void OnOffsetYChanged(double newValue)
-        {
-            UpdateShadowOffset(_dropShadow.Offset.X, (float)newValue, _dropShadow.Offset.Z);
-        }
-
-        private void OnOffsetZChanged(double newValue)
-        {
-            UpdateShadowOffset(_dropShadow.Offset.X, _dropShadow.Offset.Y, (float)newValue);
-        }
-
-        private void OnShadowOpacityChanged(double newValue)
-        {
-            _dropShadow.Opacity = (float)newValue;
-        }
-
-        private void UpdateShadowMask()
-        {
-            if (IsSupported)
-            {
-                if (Content != null)
-                {
-                    CompositionBrush mask = null;
-                    if (Content is Image)
-                    {
-                        mask = ((Image)Content).GetAlphaMask();
-                    }
-                    else if (Content is Shape)
-                    {
-                        mask = ((Shape)Content).GetAlphaMask();
-                    }
-                    else if (Content is TextBlock)
-                    {
-                        mask = ((TextBlock)Content).GetAlphaMask();
-                    }
-
-                    _dropShadow.Mask = mask;
-                }
-                else
-                {
-                    _dropShadow.Mask = null;
-                }
-            }
-        }
-
-        private void UpdateShadowOffset(float x, float y, float z)
-        {
-            _dropShadow.Offset = new Vector3(x, y, z);
-        }
-
-        private void UpdateShadowSize()
-        {
-            Vector2 newSize = new Vector2(0, 0);
-            FrameworkElement contentFE = Content as FrameworkElement;
-            if (contentFE != null)
-            {
-                newSize = new Vector2((float)contentFE.ActualWidth, (float)contentFE.ActualHeight);
-            }
-
-            _shadowVisual.Size = newSize;
-        }
-    }
-}
