+++ Microsoft.Toolkit.Uwp.UI.Controls/DropShadowPanel/DropShadowPanel.cs
@@ -10,10 +10,13 @@
 // THE CODE OR THE USE OR OTHER DEALINGS IN THE CODE.
 // ******************************************************************
 
+using System.Numerics;
+using Windows.UI;
 using Windows.UI.Composition;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Hosting;
+using Windows.UI.Xaml.Shapes;
 
 namespace Microsoft.Toolkit.Uwp.UI.Controls
 {
@@ -22,11 +25,9 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
     /// making it easier to add shadows to Xaml without having to directly drop down to Windows.UI.Composition APIs.
     /// </summary>
     [TemplatePart(Name = PartShadow, Type = typeof(Border))]
-    [TemplatePart(Name = PartContent, Type = typeof(ContentPresenter))]
     public partial class DropShadowPanel : ContentControl
     {
         private const string PartShadow = "ShadowElement";
-        private const string PartContent = "CastingElement";
 
         private readonly DropShadow _dropShadow;
         private readonly SpriteVisual _shadowVisual;
@@ -49,9 +50,9 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
                 _shadowVisual.Shadow = _dropShadow;
             }
 
-            SizeChanged += CompositionShadow_SizeChanged;
+            SizeChanged += OnSizeChanged;
 
-            Loaded += CompositionShadow_Loaded;
+            Loaded += OnLoaded;
         }
 
         /// <summary>
@@ -59,6 +60,11 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
         /// </summary>
         protected override void OnApplyTemplate()
         {
+            if (!IsSupported)
+            {
+                return;
+            }
+
             _border = GetTemplateChild(PartShadow) as Border;
 
             if (_border != null)
@@ -67,6 +73,125 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
             }
 
             base.OnApplyTemplate();
+        }
+
+        private void OnLoaded(object sender, RoutedEventArgs e)
+        {
+            ConfigureShadowVisualForCastingElement();
+        }
+
+        private void OnSizeChanged(object sender, SizeChangedEventArgs e)
+        {
+            UpdateShadowSize();
+        }
+
+        private void ConfigureShadowVisualForCastingElement()
+        {
+            UpdateShadowMask();
+
+            UpdateShadowSize();
+        }
+
+        private void OnBlurRadiusChanged(double newValue)
+        {
+            if (_dropShadow != null)
+            {
+                _dropShadow.BlurRadius = (float)newValue;
+            }
+        }
+
+        private void OnColorChanged(Color newValue)
+        {
+            if (_dropShadow != null)
+            {
+                _dropShadow.Color = newValue;
+            }
+        }
+
+        private void OnOffsetXChanged(double newValue)
+        {
+            if (_dropShadow != null)
+            {
+                UpdateShadowOffset((float)newValue, _dropShadow.Offset.Y, _dropShadow.Offset.Z);
+            }
+        }
+
+        private void OnOffsetYChanged(double newValue)
+        {
+            if (_dropShadow != null)
+            {
+                UpdateShadowOffset(_dropShadow.Offset.X, (float)newValue, _dropShadow.Offset.Z);
+            }
+        }
+
+        private void OnOffsetZChanged(double newValue)
+        {
+            if (_dropShadow != null)
+            {
+                UpdateShadowOffset(_dropShadow.Offset.X, _dropShadow.Offset.Y, (float)newValue);
+            }
+        }
+
+        private void OnShadowOpacityChanged(double newValue)
+        {
+            if (_dropShadow != null)
+            {
+                _dropShadow.Opacity = (float)newValue;
+            }
+        }
+
+        private void UpdateShadowMask()
+        {
+            if (!IsSupported)
+            {
+                return;
+            }
+
+            if (Content != null)
+            {
+                CompositionBrush mask = null;
+                if (Content is Image)
+                {
+                    mask = ((Image)Content).GetAlphaMask();
+                }
+                else if (Content is Shape)
+                {
+                    mask = ((Shape)Content).GetAlphaMask();
+                }
+                else if (Content is TextBlock)
+                {
+                    mask = ((TextBlock)Content).GetAlphaMask();
+                }
+
+                _dropShadow.Mask = mask;
+            }
+            else
+            {
+                _dropShadow.Mask = null;
+            }
+        }
+
+        private void UpdateShadowOffset(float x, float y, float z)
+        {
+            if (_dropShadow != null)
+            {
+                _dropShadow.Offset = new Vector3(x, y, z);
+            }
+        }
+
+        private void UpdateShadowSize()
+        {
+            if (_shadowVisual != null)
+            {
+                Vector2 newSize = new Vector2(0, 0);
+                FrameworkElement contentFE = Content as FrameworkElement;
+                if (contentFE != null)
+                {
+                    newSize = new Vector2((float)contentFE.ActualWidth, (float)contentFE.ActualHeight);
+                }
+
+                _shadowVisual.Size = newSize;
+            }
         }
     }
 }
